renames blockconsumer into processor
