+++ engine/verification/assigner/processor/blockjob.go
@@ -1,4 +1,4 @@
-package blockconsumer
+package processor
 
 import (
 	"fmt"
