+++ engine/verification/assigner/processor/consumer.go
@@ -1,4 +1,4 @@
-package blockconsumer
+package processor
 
 import (
 	"fmt"
