+++ engine/verification/assigner/processor/consumer_test.go
@@ -1,4 +1,4 @@
-package blockconsumer
+package processor
 
 import (
 	"fmt"
