+++ engine/verification/assigner/processor/reader.go
@@ -1,4 +1,4 @@
-package blockconsumer
+package processor
 
 import (
 	"fmt"
