+++ engine/verification/assigner/processor/reader_test.go
@@ -1,4 +1,4 @@
-package blockconsumer
+package processor
 
 import (
 	"testing"
